@@ -23,7 +23,7 @@
     "express": "^4.15.0",
     "jquery": "^3.1.1",
     "mongoose": "^5.10.15",
-    "react": "^16.4",
-    "react-dom": "^16.4"
+    "react": "^16.8.0",
+    "react-dom": "^16.8.0"
   }
 }
