@@ -1,30 +1,24 @@
-import React from 'react';
+import React, { useState } from 'react';
 
-class Search extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      term: ''
-    }
+const Search = ({ onSearch }) => {
+
+  const[term, setTerm] = useState('')
+
+  const onChange = (e) => {
+    setTerm(e.target.value);
   }
 
-  onChange (e) {
-    this.setState({
-      term: e.target.value
-    });
+  const search = () => {
+    onSearch(term);
   }
-
-  search() {
-    this.props.onSearch(this.state.term);
-  }
-
-  render() {
-    return (<div>
+//apparently there's a bug here
+  return (
+    <div>
       <h4>Add more repos!</h4>
-      Enter a github username: <input value={this.state.terms} onChange={this.onChange}/>       
-      <button onClick={this.search}> Add Repos </button>
-    </div>) 
-  }
+      Enter a github username: <input value={term} onChange={onChange}/>
+      <button onClick={search}> Add Repos </button>
+    </div>
+  );
 }
 
 export default Search;
