@@ -1,9 +1,9 @@
 import React from 'react';
 
-const RepoList = (props) => (
+const RepoList = ({ repos }) => (
   <div>
     <h4> Repo List Component </h4>
-    There are {props.repos.length} repos.
+    There are {repos.length} repos.
   </div>
 )
 
