@@ -1,30 +1,26 @@
-import React from 'react';
+import React, { useState } from 'react';
 import ReactDOM from 'react-dom';
 import $ from 'jquery';
 import Search from './components/Search.jsx';
 import RepoList from './components/RepoList.jsx';
 
-class App extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = { 
-      repos: []
-    }
+const App = () => {
 
-  }
+  const [repos, setRepos] = useState([]);
 
-  search (term) {
+  const search = (term) => {
     console.log(`${term} was searched`);
-    // TODO
+    //use axios and post
   }
+  //
 
-  render () {
-    return (<div>
+  return (
+    <div>
       <h1>Github Fetcher</h1>
-      <RepoList repos={this.state.repos}/>
-      <Search onSearch={this.search.bind(this)}/>
-    </div>)
-  }
+      <RepoList repos={repos}/>
+      <Search onSearch={search}/>
+    </div>
+  );
 }
 
 ReactDOM.render(<App />, document.getElementById('app'));
